refactor: 태그 편집 그룹 컴포넌트 분리
